pyproject.toml
[project]
name = "e-commerce-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0",
    "django-filter>=25.2",
    "django-redis>=6.0.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.11",
    "pycountry>=24.6.1",
    "redis>=7.1.0",
    "whitenoise>=6.9.0",
    "gunicorn>=23.0.0",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=6.1.0",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-django>=4.11.1",
    "pytest-watch>=4.2.0",
]
